feat: refactoring job hooks
