+++ frontend/src/components/full-screen-loader/FullScreenLoader.tsx
@@ -1,8 +1,8 @@
 import './FullScreenLoader.css';
 
-const FullScreenLoader = () => {
+const Loader = () => {
   return (
-    <div className="flex min-h-screen items-center justify-center bg-gray-800">
+    <div className="flex min-h-screen items-center justify-center">
       <div className="loader flex space-x-3 rounded-full bg-white p-5">
         <div className="h-5 w-5 animate-bounce rounded-full bg-gray-800"></div>
         <div className="h-5 w-5 animate-bounce rounded-full bg-gray-800"></div>
@@ -12,4 +12,4 @@ const FullScreenLoader = () => {
   );
 };
 
-export default FullScreenLoader;
+export default Loader;
