+++ frontend/src/components/single-contract/SingleContractData.tsx
@@ -1,13 +1,16 @@
 import DataTable from '../data-table/DataTable';
 import { IJobData } from 'interfaces/IJobData';
+import { useJob } from '../smart-contracts/hooks/useJob';
 
 interface ISingleContractDataProps {
-  contract: IJobData;
+  contract?: IJobData;
 }
 
 const SingleContractData: React.FC<ISingleContractDataProps> = (props) => {
-  const { contract } = props;
-  return <DataTable contract={contract} />;
+  const { job } = useJob();
+  const { contract = job } = props;
+
+  return contract ? <DataTable contract={contract} /> : null;
 };
 
 export default SingleContractData;
