+++ frontend/src/enums/CacheKeys.ts
@@ -1,0 +1,5 @@
+export enum CacheKeys {
+  JOB_META = 'm',
+  VERSION = 'v',
+  TIMESTAMP = 't',
+}
