+++ frontend/src/interfaces/IJobCacheEntry.ts
@@ -1,0 +1,8 @@
+import { CacheKeys } from '../enums/CacheKeys';
+import { IJobMetaData } from './IJobData';
+
+export interface IJobCacheEntry {
+  [CacheKeys.JOB_META]: IJobMetaData;
+  [CacheKeys.VERSION]: number;
+  [CacheKeys.TIMESTAMP]: number;
+}
