+++ frontend/src/pages/Contracts.tsx
@@ -1,16 +1,17 @@
 import ContractsContainer from '../components/contracts/ContractsContainer';
 import SortFilterHeading from '../components/SortFilterHeading';
-import { useFindJobs } from 'components/smart-contracts/hooks/useJob';
+import Loader from '../components/full-screen-loader/FullScreenLoader';
+import { useJobs } from '../components/smart-contracts/hooks/useJobs';
 
 const Contracts = () => {
-  const { jobs, isLoading } = useFindJobs();
+  const { jobs, isLoading } = useJobs();
 
   return (
     <>
       <SortFilterHeading heading="Contracts" displayCreate />
       {isLoading ? (
-        <div>Loading..</div>
-      ) : jobs.length ? (
+        <Loader />
+      ) : jobs?.length ? (
         <ContractsContainer jobs={jobs} />
       ) : (
         <div>No jobs found.</div>
